requirements.txt
absl-py==2.1.0
accelerate==1.10.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
albucore==0.0.21
albumentations==1.4.22
annotated-types==0.7.0
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.8
coloredlogs==15.0.1
comm==0.2.3
contourpy==1.3.1
cycler==0.12.1
Cython==3.0.11
datasets==2.21.0
debugpy==1.8.16
decorator==5.2.1
dill==0.3.8
dlib==19.24.6
docker-pycreds==0.4.0
easydict==1.13
efficientnet_pytorch==0.7.1
einops==0.8.0
eval_type_backport==0.2.0
exceptiongroup==1.3.0
executing==2.2.0
facenet-pytorch==2.5.3
filelock==3.16.1
flatbuffers==24.3.25
fonttools==4.55.3
frozenlist==1.7.0
fsspec==2024.6.1
fvcore==0.1.5.post20221221
gitdb==4.0.12
GitPython==3.1.44
grpcio==1.68.1
huggingface-hub==0.27.1
humanfriendly==10.0
idna==3.10
imageio==2.36.1
imgaug==0.4.0
insightface==0.7.3
iopath==0.1.10
ipykernel==6.30.1
ipython==8.37.0
jedi==0.19.2
Jinja2==3.1.4
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.7
kornia==0.7.4
kornia_rs==0.1.8
lazy_loader==0.4
lmdb==1.6.2
loralib==0.1.2
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mpmath==1.3.0
multidict==6.6.4
multiprocess==0.70.16
nest-asyncio==1.6.0
networkx==3.4.2
numpy==1.26.3
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
onnx==1.17.0
onnxruntime==1.20.1
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==24.2
pandas==2.2.3
parso==0.8.4
peft==0.17.0
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.6
portalocker==3.1.1
prettytable==3.12.0
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==5.29.1
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==21.0.0
pydantic==2.10.3
pydantic_core==2.27.1
Pygments==2.19.2
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyWavelets==1.8.0
PyYAML==6.0.2
pyzmq==27.0.1
regex==2024.11.6
requests==2.32.3
safetensors==0.5.0
scikit-image==0.25.0
scikit-learn==1.6.0
scipy==1.14.1
sentry-sdk==2.22.0
setproctitle==1.3.5
shapely==2.0.6
simplejson==3.19.3
simsimd==6.2.1
six==1.17.0
smmap==5.0.2
stack-data==0.6.3
stringzilla==3.11.1
sympy==1.13.1
tabulate==0.9.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
termcolor==2.5.0
threadpoolctl==3.5.0
tifffile==2024.12.12
timm==1.0.12
tokenizers==0.21.0
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.47.1
triton==3.2.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wandb==0.19.8
wcwidth==0.2.13
Werkzeug==3.1.3
xxhash==3.5.0
yacs==0.1.8
yarl==1.20.1
